Updated Map Parser, Game Manager, Spawners, Chest Models. Installed uuid
Updated Map Parser as Tiled's output has changed since the tutorial was made. level_large.json was made using old Tiled, with the new version outputting properties as an object that needs further parsing.

Set up spawner for chests. Spawns on intervals and selects an empty location.

added UUID to ChestModel.js to generate a unique ID for the entity